# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "FrequencyEnum": {
        "description": "Habit frequency label \u2014 derived from schedules.days_of_week.",
        "enum": [
          "daily",
          "weekdays",
          "custom"
        ],
        "title": "FrequencyEnum",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Habit": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "frequency": {
            "$ref": "#/components/schemas/FrequencyEnum"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_shared_habit": {
            "title": "Is Shared Habit",
            "type": "boolean"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "pending_schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleView"
              },
              {
                "type": "null"
              }
            ]
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleView"
              },
              {
                "type": "null"
              }
            ]
          },
          "schedules": {
            "items": {
              "$ref": "#/components/schemas/ScheduleView"
            },
            "title": "Schedules",
            "type": "array"
          },
          "source_challenge_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Challenge Id"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "required": [
          "name",
          "id",
          "frequency",
          "start_date",
          "is_shared_habit",
          "source_challenge_id",
          "archived_at",
          "created_at"
        ],
        "title": "Habit",
        "type": "object"
      },
      "HabitCreate": {
        "properties": {
          "days_of_week": {
            "items": {
              "type": "integer"
            },
            "maxItems": 7,
            "minItems": 1,
            "title": "Days Of Week",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "frequency": {
            "$ref": "#/components/schemas/FrequencyEnum"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "reminder_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Time"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to today (user-local); must be >= today",
            "title": "Start Date"
          },
          "target": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "required": [
          "name",
          "frequency",
          "days_of_week"
        ],
        "title": "HabitCreate",
        "type": "object"
      },
      "HabitFromSheet": {
        "additionalProperties": false,
        "properties": {
          "overrides": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HabitFromSheetOverrides"
              },
              {
                "type": "null"
              }
            ]
          },
          "sheet_habit_id": {
            "format": "uuid",
            "title": "Sheet Habit Id",
            "type": "string"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          }
        },
        "required": [
          "sheet_habit_id"
        ],
        "title": "HabitFromSheet",
        "type": "object"
      },
      "HabitFromSheetOverrides": {
        "additionalProperties": false,
        "properties": {
          "days_of_week": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "maxItems": 7,
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Of Week"
          },
          "reminder_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Time"
          }
        },
        "title": "HabitFromSheetOverrides",
        "type": "object"
      },
      "HabitList": {
        "properties": {
          "habits": {
            "items": {
              "$ref": "#/components/schemas/Habit"
            },
            "title": "Habits",
            "type": "array"
          }
        },
        "required": [
          "habits"
        ],
        "title": "HabitList",
        "type": "object"
      },
      "HabitSheetList": {
        "properties": {
          "sheets": {
            "items": {
              "$ref": "#/components/schemas/HabitSheetView"
            },
            "title": "Sheets",
            "type": "array"
          }
        },
        "required": [
          "sheets"
        ],
        "title": "HabitSheetList",
        "type": "object"
      },
      "HabitSheetView": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "habits": {
            "items": {
              "$ref": "#/components/schemas/SheetHabitView"
            },
            "title": "Habits",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "category",
          "habits"
        ],
        "title": "HabitSheetView",
        "type": "object"
      },
      "HabitUpdate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "target": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "title": "HabitUpdate",
        "type": "object"
      },
...
      "ScheduleUpdate": {
        "additionalProperties": false,
        "properties": {
          "days_of_week": {
            "items": {
              "type": "integer"
            },
            "maxItems": 7,
            "minItems": 1,
            "title": "Days Of Week",
            "type": "array"
          },
          "effective_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to tomorrow (user-local); must be > today",
            "title": "Effective From"
          },
          "frequency": {
            "$ref": "#/components/schemas/FrequencyEnum"
          },
          "reminder_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Time"
          }
        },
        "required": [
          "frequency",
          "days_of_week"
        ],
        "title": "ScheduleUpdate",
        "type": "object"
      },
      "ScheduleView": {
        "properties": {
          "days_of_week": {
            "description": "0=Mon .. 6=Sun",
            "items": {
              "type": "integer"
            },
            "maxItems": 7,
            "minItems": 1,
            "title": "Days Of Week",
            "type": "array"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "reminder_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Time"
          }
        },
        "required": [
          "days_of_week",
          "effective_from"
        ],
        "title": "ScheduleView",
        "type": "object"
      },
      "SheetHabitView": {
        "properties": {
          "days_of_week": {
            "items": {
              "type": "integer"
            },
            "title": "Days Of Week",
            "type": "array"
          },
          "default_reminder_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Reminder Time"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "frequency": {
            "$ref": "#/components/schemas/FrequencyEnum"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "frequency",
          "days_of_week",
          "default_reminder_time",
          "target"
        ],
        "title": "SheetHabitView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Habit Tracker API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/habit-sheets": {
      "get": {
        "operationId": "get_habit_sheets_api_v1_habit_sheets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitSheetList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Habit Sheets"
      }
    },
    "/api/v1/habits": {
      "get": {
        "operationId": "get_habits_api_v1_habits_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Habits"
      },
      "post": {
        "operationId": "create_habit_api_v1_habits_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HabitCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Habit"
      }
    },
    "/api/v1/habits/from-sheet": {
      "post": {
        "operationId": "create_habit_from_sheet_api_v1_habits_from_sheet_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HabitFromSheet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Habit From Sheet"
      }
    },
    "/api/v1/habits/{habit_id}": {
      "delete": {
        "operationId": "archive_habit_api_v1_habits__habit_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "habit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Habit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Archive Habit"
      },
      "get": {
        "operationId": "get_habit_api_v1_habits__habit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "habit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Habit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Habit"
      },
      "patch": {
        "operationId": "update_habit_api_v1_habits__habit_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "habit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Habit Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HabitUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Habit"
      }
    },
    "/api/v1/habits/{habit_id}/restore": {
      "post": {
        "operationId": "restore_habit_api_v1_habits__habit_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "habit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Habit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Restore Habit"
      }
    },
    "/api/v1/habits/{habit_id}/schedule": {
      "put": {
        "operationId": "update_schedule_api_v1_habits__habit_id__schedule_put",
        "parameters": [
          {
            "in": "path",
            "name": "habit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Habit Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Schedule"
      }
    },
...
  }
}